Moved ModelPolicyStateMaps to the defaultModelChains file.
